Fix spell check failure by using en-US variant of word
For the sake of consistency, the codespell tool used to check spelling in the project files is configured to use en-US spellings.

The word licenced was added to the tool's dictionary of non-en-US variants in the latest release. This addition caused the spell check to begin to fail since the word is present in the project's files.

The chosen solution is to change the to using the en-US variant of the word.